Removed NAT, fixed missing moles, removed singles
This change adds in a lot of bug fixes and fixes many things.
(1) Removing anyone with frac_complete < .05 removed some moles without removing their imposter. This left an isolated imposter (master) with no mole. Additionally, its possible to have frac_complete < .001 and still have many show_answer in some courses! So I removed this completely.
(2) Instead we remove any mole who has too many percent_correct, to few nshow_answer (didn't even harvest much), or never did anything (frac_complete = 0) via: "certified = false and (percent_correct_attempts > 70 or nshow_answer <= 5 or frac_complete = 0) as remove"
(3) We realize that removing these users could remove all the moles, in which case there is no mole and so the imposter is in fact a valid student and should not be included. Thus, I added an outer select that removes any imposters with no moles or vice versa.
(4) I also added a column for country which has been very useful for understanding the data.
(5) Lastly, I added ipcnt < 8 which is a very very rough way to remove the case when a bunch of students are working on a NAT together. If a user has 7 moles, plus their imposter account, then we end up filtering them out, but such a case is not really a viable strategy for a student.